requirements.txt
# requirements.txt
google-cloud-storage==2.18.2
google-genai==0.2.2
requests==2.31.0
google-maps-addressvalidation